fix(#1015): extract PluginBuilder ability classes into PSR-4 files and add partition_tool_calls() to AgentLoop
- GeneratePluginAbility, SandboxTestPluginAbility, SandboxActivatePluginAbility,
  UpdatePluginSandboxedAbility, ScanPluginHooksAbility, ScanThemeHooksAbility each
  get their own file under includes/Abilities/ so PSR-4 autoloading can resolve them.
- PluginBuilderAbilities.php is trimmed to just the registry class; ability class
  definitions are removed (they now live in their own files).
- AgentLoop::partition_tool_calls() private method added (delegates to
  ClientAbilityRouter::partition()). The inline call is updated to use it.
  Tests exercise this via ReflectionClass.

Resolves #1015
